Removed fake home power calculation
It makes absolute sense, that the sum of powers is not exactly zero, due to conversion losses. These losses are valueable information and should not be discarded by a communication library like this.

If one wants to use "prettified" numbers later on, this calculation should happen outside of the communication library.